# apps/delivery-api/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CLEVER Route Server Delivery API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  summary: CLEVER route server delivery API HTTP contract
  description: >-
    Machine-readable contract for CLEVER route server communication with current
    admin clients, legacy Shopify compatibility routes, native driver mobile
    clients, and operational health checks. Human behavior notes live beside
    this file in docs/api/*.md.
  x-contract-revision: 2026-05-14
  x-source-repository: EVNSolution/clever-route-server
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://{deliveryServerHost}
    description: Deployed delivery server
    variables:
      deliveryServerHost:
        default: delivery.example.com
security: []
...
paths:
...
  /admin/drivers:
    get:
      tags: [Shopify Embedded Admin]
      summary: List drivers for the authenticated shop
      operationId: listAdminDrivers
      x-source-file: src/routes/admin-drivers.routes.ts
      security:
        - ShopifySessionToken: []
      responses:
        '200':
          description: Drivers for the authenticated shop.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDriversListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Shopify Embedded Admin]
      summary: Create a pending driver for the authenticated shop
      operationId: createAdminDriver
      x-source-file: src/routes/admin-drivers.routes.ts
      security:
        - ShopifySessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDriverCreateRequest'
      responses:
        '201':
          description: Driver was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDriverEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ShopifySessionToken:
      type: http
      scheme: bearer
      bearerFormat: Shopify App Bridge session JWT
      description: Server-verified Shopify embedded app session token.
    DriverAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-issued short-lived driver JWT with audience clever-delivery-driver.
    ShopifyWebhookHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
      description: Base64 HMAC generated by Shopify over the raw request body.
...
  responses:
    BadRequest:
      description: Invalid request payload, query, path, or headers.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid bearer token or HMAC.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/Error'
...
    AdminDriverCreateRequest:
      type: object
      required: [source, phone]
      properties:
        source:
          const: clever-app-driver-invite
        phone:
          pattern: '^\+[1-9]\d{6,14}$'
          type: string
        displayName:
          type:
            - string
            - 'null'
        inviteLink:
          type:
            - string
            - 'null'
          format: uri
      additionalProperties: false
    AdminDriver:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        displayName:
          type:
            - string
            - 'null'
      additionalProperties: true
    AdminDriverEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [driver]
          properties:
            driver:
              $ref: '#/components/schemas/AdminDriver'
        error:
          type: 'null'
    AdminDriversListEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [drivers]
          properties:
            drivers:
              type: array
              items:
                $ref: '#/components/schemas/AdminDriver'
        error:
          type: 'null'
...